[14.0][FIX] l10n_it_declaration_of_intent spostato parzialmente migrazione in pre-migrate migration
